iam-gateway/requirements.txt
flask==2.2.3
PyJWT==2.6.0
gunicorn==20.1.0
requests==2.28.2